requirements.txt
# Python Backend Dependencies
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Redis
redis==5.0.1
redis-om==0.2.1

# Kafka
kafka-python==2.0.2
aiokafka==0.10.0

# Machine Learning & Statistics
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
pymc==5.10.4
arviz==0.17.0
scipy==1.11.4
statsmodels==0.14.1

# Deep Learning (Optional)
tensorflow==2.15.0
torch==2.1.2

# NLP & Embeddings
sentence-transformers==2.3.1
transformers==4.36.2
chromadb==0.4.22

# LLM Providers
anthropic==0.8.1
openai==1.10.0

# Computer Vision
Pillow==10.2.0
opencv-python==4.9.0.80

# Data Processing
polars==0.20.3
duckdb==0.10.0
pyarrow==14.0.2

# Workflow & Scheduling
prefect==2.14.21
celery[redis]==5.3.6
apscheduler==3.10.4

# HTTP & API Clients
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.0
pyjwt==2.8.0

# Validation & Serialization
marshmallow==3.20.1
pydantic-extra-types==2.4.1
email-validator==2.1.0

# Observability
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
sentry-sdk[fastapi]==1.39.2

# Logging
python-json-logger==2.0.7
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0
faker==22.2.0
factory-boy==3.3.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pylint==3.0.3

# Environment & Config
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2

# Utilities
python-slugify==8.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
pendulum==3.0.0
arrow==1.3.0

# File Processing
openpyxl==3.1.2
xlsxwriter==3.1.9
python-docx==1.1.0
pypdf2==3.0.1

# CLI
click==8.1.7
rich==13.7.0
typer==0.9.0

# Development Tools
ipython==8.20.0
jupyter==1.0.0
notebook==7.0.6

# Data Visualization (for notebooks)
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# External API Clients
google-cloud-bigquery==3.14.1
google-cloud-storage==2.14.0
google-analytics-data==0.18.5
shopify-python-api==12.3.0
tweepy==4.14.0
praw==7.7.1

# dbt
dbt-core==1.7.4
dbt-postgres==1.7.4
dbt-bigquery==1.7.2

# Type Stubs
types-redis==4.6.0.11
types-requests==2.31.0.20
types-PyYAML==6.0.12.12

# Performance
cython==3.0.8
ujson==5.9.0
orjson==3.9.12
